Add unity build for serval modules
