+++ bcos-executor/CMakeLists.txt
@@ -15,6 +15,7 @@ find_package(Boost REQUIRED serialization program_options context)
 find_package(jsoncpp REQUIRED)
 
 file(GLOB_RECURSE SRCS src/*.cpp)
+message("All executor files is: ${SRCS}")
 add_library(${EXECUTOR_TARGET} ${SRCS})
 if(WITH_WASM)
     target_link_libraries(${EXECUTOR_TARGET} PUBLIC jsoncpp_static ${CODEC_TARGET} ${TOOL_TARGET} ${CRYPTO_TARGET} ${TABLE_TARGET} wedprcrypto::fisco-bcos bcos-protocol
@@ -23,6 +24,10 @@ else()
     target_link_libraries(${EXECUTOR_TARGET} PUBLIC jsoncpp_static ${CODEC_TARGET} ${TOOL_TARGET} ${CRYPTO_TARGET} ${TABLE_TARGET} wedprcrypto::fisco-bcos bcos-protocol
             Boost::context evmone evmc::loader evmc::instructions GroupSig Paillier blst)
 endif()
+# set_target_properties(${EXECUTOR_TARGET} PROPERTIES
+#     UNITY_BUILD "ON"
+# )
+# set(SKIP_UNITY_BUILD_INCLUSION "src/vm/")
 
 if (TOOLS)
     add_subdirectory(tools)
