+++ bcos-framework/bcos-framework/protocol/ProtocolTypeDef.h
@@ -80,25 +80,6 @@ using WeightListPtr = std::shared_ptr<WeightList>;
 
 int64_t constexpr InvalidSealerIndex = INT64_MAX;
 
-struct Session
-{
-    using Ptr = std::shared_ptr<Session>;
-    using ConstPtr = std::shared_ptr<const Session>;
-
-    enum Status
-    {
-        STARTED = 0,
-        DIRTY,
-        COMMITTED,
-        ROLLBACKED
-    };
-
-    long sessionID;
-    Status status;
-    bcos::protocol::BlockNumber beginNumber;  // [
-    bcos::protocol::BlockNumber endNumber;    // )
-};
-
 struct TwoPCParams
 {
     BlockNumber number = 0;
