+++ bcos-gateway/CMakeLists.txt
@@ -29,6 +29,9 @@ find_package(tarscpp REQUIRED)
 
 add_library(${GATEWAY_TARGET} ${SRCS})
 target_link_libraries(${GATEWAY_TARGET} PUBLIC jsoncpp_static Boost::filesystem bcos-boostssl ${FRONT_TARGET} ${TARS_PROTOCOL_TARGET} tarscpp::tarsservant tarscpp::tarsutil)
+set_target_properties(${GATEWAY_TARGET} PROPERTIES
+    UNITY_BUILD "ON"
+)
 
 if (APPLE)
 # target_compile_options(${GATEWAY_TARGET} PRIVATE -faligned-allocation)
