+++ bcos-rpc/CMakeLists.txt
@@ -30,6 +30,9 @@ find_package(tarscpp REQUIRED)
 
 add_library(${RPC_TARGET} ${SRCS} ${HEADERS})
 target_link_libraries(${RPC_TARGET} PUBLIC bcos-boostssl ${LEDGER_TARGET} ${CRYPTO_TARGET} ${TARS_PROTOCOL_TARGET} jsoncpp_static tarscpp::tarsservant tarscpp::tarsutil)
+set_target_properties(${RPC_TARGET} PROPERTIES
+    UNITY_BUILD "ON"
+)
 
  if (TESTS)
     enable_testing()
