+++ bcos-scheduler/CMakeLists.txt
@@ -7,6 +7,9 @@ aux_source_directory(src SRCS)
 
 add_library(${SCHEDULER_TARGET} ${SRCS})
 target_link_libraries(${SCHEDULER_TARGET} PUBLIC ${TABLE_TARGET} ${TARS_PROTOCOL_TARGET})
+set_target_properties(${SCHEDULER_TARGET} PROPERTIES
+    UNITY_BUILD "ON"
+)
 
 if(TESTS)
     enable_testing()
