+++ bcos-tars-protocol/CMakeLists.txt
@@ -40,6 +40,9 @@ target_include_directories(${TARS_PROTOCOL_TARGET} PUBLIC
     $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/generated>
     $<INSTALL_INTERFACE:include/bcos-tars-protocol>)
 target_link_libraries(${TARS_PROTOCOL_TARGET} PUBLIC bcos-concepts bcos-crypto tarscpp::tarsservant tarscpp::tarsutil TBB::tbb)
+set_target_properties(${TARS_PROTOCOL_TARGET} PROPERTIES
+    UNITY_BUILD "ON"
+)
 
 if(TESTS)
     enable_testing()
