+++ bcos-tool/CMakeLists.txt
@@ -32,6 +32,9 @@ target_include_directories(${TOOL_TARGET} PUBLIC
     $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/bcos-protocol>
 )
 target_link_libraries(${TOOL_TARGET} PUBLIC ${TABLE_TARGET} ${UTILITIES_TARGET} bcos-framework jsoncpp_static tarscpp::tarsservant tarscpp::tarsutil TBB::tbb)
+set_target_properties(${TOOL_TARGET} PROPERTIES
+    UNITY_BUILD "ON"
+)
 
 if(TESTS)
   enable_testing()
