+++ bcos-utilities/CMakeLists.txt
@@ -25,6 +25,9 @@ if (ONLY_CPP_SDK)
 else ()
     target_link_libraries(bcos-utilities PUBLIC Boost::iostreams Boost::log Boost::filesystem Boost::chrono Boost::thread Boost::serialization redis++::redis++_static ZLIB::ZLIB zstd::libzstd_static ittapi)
 endif ()
+set_target_properties(bcos-utilities PROPERTIES
+    UNITY_BUILD "ON"
+)
 
 if(TESTS)
     enable_testing()
