+++ fisco-bcos-air/AirNodeInitializer.cpp
@@ -30,11 +30,9 @@
 #include <bcos-rpc/RpcFactory.h>
 #include <bcos-rpc/groupmgr/NodeService.h>
 #include <bcos-rpc/tarsRPC/RPCServer.h>
-#include <bcos-scheduler/src/SchedulerImpl.h>
 #include <bcos-tars-protocol/protocol/ProtocolInfoCodecImpl.h>
 #include <bcos-tool/NodeConfig.h>
 #include <rocksdb/env.h>
-#include <iterator>
 
 using namespace bcos::node;
 using namespace bcos::initializer;
