+++ transaction-scheduler/CMakeLists.txt
@@ -7,6 +7,9 @@ target_include_directories(transaction-scheduler PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
     $<INSTALL_INTERFACE:include/bcos-transaction-scheduler>)
 target_link_libraries(transaction-scheduler PUBLIC bcos-framework fmt::fmt-header-only)
+set_target_properties(transaction-scheduler PROPERTIES
+    UNITY_BUILD "ON"
+)
 
 if (TESTS)
     enable_testing()
